Functions

PolaronMobility.OldPolaron — Type

Polaron. Structure to store data of polaron solution and other parameters, for each temperature or frequency.

PolaronMobility.A — Method

A(v, w, β)

Hellwarth's A expression from Eqn. (62b) in Hellwarth et al. 1999 PRB. Part of the overall free energy expression.

See Hellwarth et a. 1999: https://doi.org/10.1103/PhysRevB.60.299.

PolaronMobility.A — Method

A(v, w, β)

Generalisation of the A function from Eqn. (62b) in Hellwarth et al. 1999. This is the Helmholtz free energy of the trial model.

Required for calculating the polaron free energy.

Arguments

v::Vector{Float64}: is a vector of the v variational parameters.

w::Vector{Float64}: is a vector of the w variational parameters.

β::Union{Float64, Vector{Float64}}: is the reduced thermodynamic temperature ħωⱼ/(kT) associated with the 'jth' phonon mode.

See Hellwarth, R. W., Biaggio, I. (1999): https://doi.org/10.1103/PhysRevB.60.299.

PolaronMobility.A — Method

A(v, w, n)

Calculates A(v, w, β) but at zero-temperature, β = Inf.

Arguments

v::Vector{Float64}: is a vector of the v variational parameters.

w::Vector{Float64}: is a vector of the w variational parameters.

PolaronMobility.B — Method

B(v, w, β, α)

Hellwarth's B expression from Eqn. (62c) in Hellwarth et al. 1999 PRB. Part of the overall free energy expression.

See Hellwarth et a. 1999: https://doi.org/10.1103/PhysRevB.60.299.

PolaronMobility.B — Method

B(v, w, α)

Integral of Eqn. (31) in Feynman 1955. Part of the overall ground-state energy expression.

See Feynman 1955: http://dx.doi.org/10.1103/PhysRev.97.660.

PolaronMobility.B — Method

B(v, w, α, β)

Generalisation of the B function from Eqn. (62c) in Hellwarth et al. 1999. This is the expected value of the exact action <S_j> taken w.r.t trial action, given for the 'jth' phonon mode.

Required for calculating the polaron free energy.

Arguments

v::Vector{Float64}: is a vector of the v variational parameters.

w::Vector{Float64}: is a vector of the w variational parameters.

α::Union{Float64, Vector{Float64}}: is the partial dielectric electron-phonon coupling parameter for the 'jth' phonon mode.

β::Union{Float64, Vector{Float64}}: is the reduced thermodynamic temperature ħωⱼ/(kT) associated with the 'jth' phonon mode.

See Hellwarth, R. W., Biaggio, I. (1999): https://doi.org/10.1103/PhysRevB.60.299.

PolaronMobility.B — Method

B(v, w, α; rtol = 1e-3)

Calculates B(v, w, α, β) but at zero-temperature, β = Inf.

Arguments

v::Vector{Float64}: is a vector of the v variational parameters.

w::Vector{Float64}: is a vector of the w variational parameters.

α::Union{Float64, Vector{Float64}}: is the partial dielectric electron-phonon coupling parameter for the 'jth' phonon mode.

PolaronMobility.B_integrand — Method

B(τ, v, w)

Integrand of Eqn. (31) in Feynman 1955. Part of the overall ground-state energy expression.

See Feynman 1955: http://dx.doi.org/10.1103/PhysRev.97.660.

PolaronMobility.C — Method

C(v, w, β)

Hellwarth's C expression from Eqn. (62e) in Hellwarth et al. 1999 PRB. Part of the overall free energy expression.

See Hellwarth et a. 1999: https://doi.org/10.1103/PhysRevB.60.299.

PolaronMobility.C — Method

C(v, w, β)

Generalisation of the C function from Eqn. (62e) in Hellwarth et al. 1999. This is the expected value of the trial action <S_0> taken w.r.t trial action.

Required for calculating the polaron free energy.

Arguments

v::Vector{Float64}: is a vector of the v variational parameters.

w::Vector{Float64}: is a vector of the w variational parameters.

β::Union{Float64, Vector{Float64}}: is the reduced thermodynamic temperature ħωⱼ/(kT) associated with the 'jth' phonon mode.

See Hellwarth, R. W., Biaggio, I. (1999): https://doi.org/10.1103/PhysRevB.60.299.

PolaronMobility.C — Method

C(v, w)

Calculates C(v, w, β) but at zero-temperature, β = Inf.

Arguments

v::Vector{Float64}: is a vector of the v variational parameters.

w::Vector{Float64}: is a vector of the w variational parameters.

PolaronMobility.C_ij — Method

C_ij(i, j, v, w)

Calculates the element to the coupling matrix C_ij (a generalisation of Feynman's C coupling variational parameter in Feynman 1955) between the electron and the ith' andjth' fictitious masses that approximates the exact electron-phonon interaction with a harmonic coupling to a massive fictitious particle.

Required for calculating the polaron free energy.

Note: Not to be confused with the number of physical phonon branches; many phonon branches could be approximated with one or two etc. fictitious masses for example. The number of fictitious mass does not necessarily need to match the number of phonon branches.

Arguments

i::Integer, j::Integer: enumerate the current fictitious masses under focus (also the index of the element in the coupling matrix C)

v::Vector{Float64}: is a vector of the v variational parameters.

w::Vector{Float64}: is a vector of the w variational parameters.

See Feynman 1955: http://dx.doi.org/10.1103/PhysRev.97.660.

PolaronMobility.D — Method

D(τ, v, w, β)

Calculates the recoil function (a generalisation of D(u) in Eqn. (35c) in FHIP 1962) that approximates the exact influence (recoil effects) of the phonon bath on the electron with the influence of the fictitious masses attached by springs to the electron. It appears in the exponent of the intermediate scattering function.

Arguments

τ::Float64: is the imaginary time variable.

v::Vector{Float64}: is a vector of the v variational parameters.

w::Vector{Float64}: is a vector of the w variational parameters.

β::Union{Float64, Vector{Float64}}: is the reduced thermodynamic temperature ħωⱼ/(kT) associated with the 'jth' phonon mode.

See FHIP 1962: https://doi.org/10.1103/PhysRev.127.1004.

PolaronMobility.D — Method

D(τ, v, w)

Calculates the recoil function at zero-temperature.

Arguments

τ::Float64: is the imaginary time variable.

v::Vector{Float64}: is a vector of the v variational parameters.

w::Vector{Float64}: is a vector of the w variational parameters.

PolaronMobility.DielectricFromIRmode — Method

DieletricFromIRmode(IRmode)

Calculate dielectric from an individual mode.

IRmode is a tuple f, S with Frequency in THz; InfraRed activity in e^2 amu^-1.

PolaronMobility.F — Method

F(v, w, α, ω)

Calculates the zero-temperature ground-state energy of the polaron for a material with multiple phonon branches. Similar to F(v, w, α, ω, β) but with β = Inf. Generalises Eqn. (33) in Feynman 1955.

Arguments

v::Vector{Float64}: is a vector of the v variational parameters.

w::Vector{Float64}: is a vector of the w variational parameters.

α::Union{Float64, Vector{Float64}}: is the partial dielectric electron-phonon coupling parameter for the 'jth' phonon mode.

ω::Union{Float64, Vector{Float64}}: phonon mode frequencies (2π THz).

See Feynman 1955: http://dx.doi.org/10.1103/PhysRev.97.660.

PolaronMobility.F — Method

F(v, w, α, ω, β)

Calculates the Helmholtz free energy of the polaron for a material with multiple phonon branches.

This generalises the Osaka 1959 (below Eqn. (22)) and Hellwarth. et al 1999 (Eqn. (62a)) free energy expressions.

Arguments

v::Vector{Float64}: is a vector of the v variational parameters.

w::Vector{Float64}: is a vector of the w variational parameters.

α::Union{Float64, Vector{Float64}}: is the partial dielectric electron-phonon coupling parameter for the 'jth' phonon mode.

ω::Union{Float64, Vector{Float64}}: phonon mode frequencies (2π THz).

β::Union{Float64, Vector{Float64}}: is the reduced thermodynamic temperature ħωⱼ/(kT) associated with the 'jth' phonon mode.

See Osaka, Y. (1959): https://doi.org/10.1143/ptp.22.437 and Hellwarth, R. W., Biaggio, I. (1999): https://doi.org/10.1103/PhysRevB.60.299.

PolaronMobility.Hellwarth1999mobilityRHS — Method

Hellwarth1999mobilityRHS((α, (v, w) ,f), effectivemass, T)

Calculates the DC mobility using Hellwarth et al. 1999 Eqn. (2).

See Hellwarth et a. 1999: https://doi.org/10.1103/PhysRevB.60.299.

PolaronMobility.HellwarthAScheme — Method

HellwarthAScheme(phonon_modes; T = 295, convergence = 1e-6)

Multiple phonon mode reduction to a single effective frequency. Temperature dependent, defaults to T = 295 K.

Solved iteratively by bisection until Δfreq < convergence.

Follows Hellwarth et al. 1999 PRB 'A' scheme, Eqn. (50) RHS.

See Hellwarth et a. 1999: https://doi.org/10.1103/PhysRevB.60.299.

PolaronMobility.HellwarthBScheme — Method

HellwarthBScheme(LO)

LO an array assumed to be of [freq ; absolute ir activity ]

Multiple phonon mode reduction to a single effective frequency. Hellwarth et al. 1999 PRB, 'B scheme'; the athermal method. Averaging procedure is constructed by considering the average effect of the action of multiple branches.

Follows Eqn. (58) in this paper, assuming typo on LHS, should actually be W_e.

See Hellwarth et a. 1999: https://doi.org/10.1103/PhysRevB.60.299.

PolaronMobility.IRtoDielectric — Method

IRtoDielectric(IRmodes, volume)

From absolute value of IR activities of phonon modes, generate a per-mode contribution to the low-frequency dielectric constant.

IRmodes are tuples f, S with Frequency in THz; InfraRed activity in e^2 amu^-1.

PolaronMobility.IRtoalpha — Method

IRtoalpha(IR, volume)

Calculates contribution to dielectric constant for each polar phonon mode, and thereby the Frohlich alpha contribution for this mode.

PolaronMobility.ImX — Method

function ImX(nurange,v,w,βred,α,ω,mb)

Impedance in (47a) from Feynman1962, directly solving freq dep without taking Hellwarth1999 limit of v->0 .

Calculates a frequency dependent (over range of nu) susceptibility which can be linked back to mobility.

HERE BE DRAGONS! Not well tested or validated code; the central numeric integration is highly oscillatory and becomes intractable for large nu.

PolaronMobility.Y — Method

Y(x, v, β)

Hellwarth's Y expression from Eqn. (62d) in Hellwarth et al. 1999 PRB. Part of the overall free energy expression. Contained in denominator of the integrand of Eqn. (62c).

See Hellwarth et a. 1999: https://doi.org/10.1103/PhysRevB.60.299.

PolaronMobility.effective_freqs — Method

effective_freqs(freqs_and_ir_activity, num_var_params)

Generates a matrix of effective phonon modes with frequencies and infra-red activities derived from a larger matrix using the Principal Component Analysis (PCA) method.

Arguments

freqs_and_ir_activity::Matrix{Float64}: is a matrix containing the phonon mode frequencies (in THz) in the first column and the infra-red activities (in e²amu⁻¹) in the second column.

num_var_params::Integer: is the number of effective modes required (which needs to be less than the number of modes in freqs_and_ir_activity`).

*** POSSIBLY REDUNDANT ***

PolaronMobility.f — Method

f(x, v, w, β)

Integrand of Hellwarth's B expression from Eqn. (62c) in Hellwarth et al. 1999 PRB. Part of the overall free energy expression.

See Hellwarth et a. 1999: https://doi.org/10.1103/PhysRevB.60.299.

PolaronMobility.feynmanvw — Method

feynmanvw(v, w, αωβ...; upper_limit = Inf64)

Minimises the multiple phonon mode free energy function for a set of vₚ and wₚ variational parameters. The variational parameters follow the inequality: v₁ > w₁ > v₂ > w₂ > ... > vₙ > wₙ. Generalises feynmanvw to multiple variational parameters.

Arguments

v::Vector{Float64}: vector of initial v parameters.

w::Vector{Float64}: vector of initial w parameters.

α::Union{Float64, Vector{Float64}}: is the partial dielectric electron-phonon coupling parameter for one or many phonon modes.

ω::Union{Float64, Vector{Float64}}: phonon mode frequencies (2π THz) for one or many phonon modes.

β::Union{Float64, Vector{Float64}}: is the reduced thermodynamic temperature ħωⱼ/(kT) for one or many phonon modes.

See also F.

PolaronMobility.frohlichPartial — Method

frohlichPartial((f, ϵ_mode); ϵ_o, ϵ_s, meff)

Calculate a (partial) dielectric electron-phonon coupling element.

Arguments

f: frequency of mode in THz.

ϵ_mode: this mode's contribution to dielectric.

ϵ_o: optical dielectric.

ϵ_s: total static dielectric contribution.

PolaronMobility.frohlichalpha — Method

frohlichalpha(ε_Inf, ε_S, freq, m_eff)

Calculates the Frohlich alpha parameter, for a given dielectric constant, frequency (f) of phonon in Hertz, and effective mass (in units of the bare electron mass).

See Feynman 1955: http://dx.doi.org/10.1103/PhysRev.97.660.

PolaronMobility.frohlichalpha — Method

frohlichalpha(ϵ_optic, ϵ_ionic, ϵ_total, phonon_mode_freq, m_eff)

Calculates the partial dielectric electron-phonon coupling parameter for a given longitudinal optical phonon mode.

This decomposes the original Frohlich alpha coupling parameter (defined for a single phonon branch) into contributions from multiple phonon branches.

Arguments

ϵ_optic::Float64: is the optical dielectric constant of the material.

ϵ_ionic::Float64: is the ionic dielectric contribution from the phonon mode.

ϵ_total::Float64: is the total ionic dielectric contribution from all phonon modes of the material.

phonon_mode_freq::Float64: is the frequency of the phonon mode (THz).

m_eff::Float64 is the band mass of the electron (in units of electron mass m_e).

PolaronMobility.h_i — Method

h_i(i, v, w)

Calculates the normal-mode (the eigenmodes) frequency of the coupling between the electron and the `ith' fictitious mass that approximates the exact electron-phonon interaction with a harmonic coupling to a massive fictitious particle.

Required for calculating the polaron free energy.

Note: Not to be confused with the number of physical phonon branches; many phonon branches could be approximated with one or two etc. fictitious masses for example. The number of fictitious mass does not necessarily need to match the number of phonon branches.

Arguments

i::Integer: enumerates the current fictitious mass.

v::Vector{Float64}: is a vector of the v variational parameters.

w::Vector{Float64}: is a vector of the w variational parameters.

PolaronMobility.load_polaron — Method

load_polaron(p::NewPolaron, prefix)

Loads data from file "polaronfilepath" into a NewPolaron type.

PolaronMobility.optical_absorption — Method

optical_absorption(Ω, β, α, v, w; rtol = 1e-3)

Calculate the absorption coefficient Γ(Ω) for the polaron at at finite temperatures (Eqn. (11a) in DSG 1972) for a given frequency Ω.

Arguments

Ω::Float64: is the frequency (2π THz) of applied electric field.

β::Float64: is the reduced thermodynamic betas.

α::Float64: is the Frohlich alpha coupling parameter.

v::Float64: is the 'v' variational parameter.

w::Float64: is the 'w' variational parameter.

rtol: relative tolerance for the accuracy of any quadrature integrations.

See DSG 1972: https://doi.org/10.1103/PhysRevB.5.2367.

PolaronMobility.polaron_complex_conductivity — Method

polaron_complex_conductivity(Ω, β, α, v, w; rtol = 1e-3)

Calculate the complex conductivity σ(Ω) of the polaron at finite temperatures for a given frequency Ω (equal to 1 / Z(Ω) with Z the complex impedence).

Arguments

Ω::Float64: is the frequency (2π THz) of applied electric field.

β::Float64: is the reduced thermodynamic betas.

α::Float64: is the Frohlich alpha coupling parameter.

v::Float64: is the 'v' variational parameter.

w::Float64: is the 'w' variational parameter.

rtol: relative tolerance for the accuracy of any quadrature integrations.

See also polaron_complex_impedence

PolaronMobility.polaron_complex_impedence — Method

polaron_complex_impedence(Ω, β, α, v, w; rtol = 1e-3, T = nothing, verbose = false)

Calculate the complex impedence Z(Ω) of the polaron at finite temperatures for a given frequency Ω (Eqn. (41) in FHIP 1962).

Arguments

Ω::Float64: is the frequency (2π THz) of applied electric field.

β::Float64: is the reduced thermodynamic betas.

α::Float64: is the Frohlich alpha coupling parameter.

v::Float64: is the 'v' variational parameter.

w::Float64: is the 'w' variational parameter.

rtol: relative tolerance for the accuracy of any quadrature integrations.

T: is a token used by make_polaron() to keep track of the temperature for printing during a calculation. Do not alter.

verbose: is used by make_polaron() to specify whether or not to print. Ignore.

See FHIP 1962: https://doi.org/10.1103/PhysRev.127.1004.

PolaronMobility.polaron_memory_function — Method

polaron_memory_function(Ω, β, α, v, w; ω = 1.0, rtol = 1e-3)

Calculate the memory function χ(Ω) of the polaron at finite temperatures (equation (35) in FHIP 1962) for a given frequency Ω.

Includes conditionals for zero-temperature and DC limits.

Arguments

Ω::Float64: electric field frequency (2π THz).

β::Union{Float64, Vector{Float64} reduced thermodynamic temperatures (unitless). Either a single value for one phonon frequency or a vector of values for multiple phonon frequencies.

α::Union{Float64, Vector{Float64}}: Frohlich alpha coupling parameter (unitless). Either a single value for one phonon frequency or a vector of values for multiple phonon frequencies.

v::Union{Float64, Vector{Float64}}: variational 'v' parameters that minimise the polaron free energy (unitless).

w::Union{Float64, Vector{Float64}}: variational 'w' parameters that minimise the polaron free energy (unitless).

ω::Union{Float64, Vector{Float64}}: phonon mode frequencies (2π THz). Predefined as ω = 1.0 for a single mode in polaron units.

rtol = 1e-3: specifies the relative error tolerance for the QuadGK integral.

See R. P. Feynman, R. W. Hellwarth, C. K. Iddings, and P. M. Platzman, Mobility of slow electrons in a polar crystal, PhysicalReview127, 1004 (1962). https://doi.org/10.1103/PhysRev.127.1004.

See also polaron_memory_function_thermal, polaron_memory_function_athermal, polaron_memory_function_dc.

PolaronMobility.polaron_memory_function_athermal — Method

function polaron_memory_function_athermal(Ω, α, v, w; ω = 1.0, rtol = 1e-3)

Calculate polaron complex memory function at zero-temperature inclusive of multiple phonon branches j, each with angular frequency ωj.

Arguments

Ω::Float64: is the frequency (THz) of applied electric field.

α::Vector{Float64}: is an array of decomposed Frohlich alphas, one for each phonon frequency ω[j].

v::Union{Float64, Vector{Float64}}: is a vector or single value of the 'v' variational parameter(s).

w::Union{Float64, Vector{Float64}}: is a vector or single value of the 'w' variational parameter(s).

m_eff::Float64: is the bare-band mass of the particle (typically electron / hole, in units of electron mass m_e).

ω::Union{Float64, Vector{Float64}}: phonon mode frequencies (2π THz). Predefined as ω = 1.0 for a single mode in polaron units.

rtol = 1e-3: specifies the relative error tolerance for the QuadGK integral.

See also polaron_memory_function.

PolaronMobility.polaron_memory_function_dc — Method

function polaron_memory_function_dc(β, α, v, w; ω = 1.0, rtol = 1e-3)

Calculate zero-frequency polaron complex memory function at finite temperature inclusive of multiple phonon branches j, each with angular frequency ωj.

Arguments

β::Vector{Float64}: is an array of reduced thermodynamic betas, one for each phonon frequency ω[j].

α::Vector{Float64}: is an array of decomposed Frohlich alphas, one for each phonon frequency ω[j].

v::Union{Float64, Vector{Float64}}: is a vector or single value of the 'v' variational parameter(s).

w::Union{Float64, Vector{Float64}}: is a vector or single value of the 'w' variational parameter(s).

m_eff::Float64: is the bare-band mass of the particle (typically electron / hole, in units of electron mass m_e).

ω::Union{Float64, Vector{Float64}}: phonon mode frequencies (2π THz). Predefined as ω = 1.0 for a single mode in polaron units.

rtol = 1e-3: specifies the relative error tolerance for the QuadGK integral.

See also polaron_memory_function.

PolaronMobility.polaron_memory_function_thermal — Method

function polaron_memory_function_thermal(Ω, β, α, v, w; ω = 1.0, rtol = 1e-3)

Calculate polaron complex memory function at finite temperature inclusive of multiple phonon branches j, each with angular frequency ωj.

Arguments

Ω::Float64: is the frequency (THz) of applied electric field.

β::Vector{Float64}: is an array of reduced thermodynamic betas, one for each phonon frequency ω[j].

α::Vector{Float64}: is an array of decomposed Frohlich alphas, one for each phonon frequency ω[j].

v::Union{Float64, Vector{Float64}}: is a vector or single value of the 'v' variational parameter(s).

w::Union{Float64, Vector{Float64}}: is a vector or single value of the 'w' variational parameter(s).

m_eff::Float64: is the bare-band mass of the particle (typically electron / hole, in units of electron mass m_e).

ω::Union{Float64, Vector{Float64}}: phonon mode frequencies (2π THz). Predefined as ω = 1.0 for a single mode in polaron units.

rtol = 1e-3: specifies the relative error tolerance for the QuadGK integral.

See also polaron_memory_function.

PolaronMobility.polaron_mobility — Method

polaron_mobility(β, α, v, w; rtol = 1e-3, T = nothing, verbose = false)

Calculate the dc mobility μ of the polaron at finite temperatues (Eqn. (11.5) in F. Peeters and J. Devreese 1984) for a given frequency Ω.

Arguments

β::Float64: is the reduced thermodynamic betas.

α::Float64: is the Frohlich alpha coupling parameter.

v::Float64: is the 'v' variational parameter.

w::Float64: is the 'w' variational parameter.

rtol: relative tolerance for the accuracy of any quadrature integrations.

T: is a token used by make_polaron() to keep track of the temperature for printing during a calculation. Do not alter.

verbose: is used by make_polaron() to specify whether or not to print. Ignore.

See F. Peeters and J. Devreese 1984: https://doi.org/10.1016/S0081-1947(08)60312-4.

See also polaron_complex_conductivity

PolaronMobility.polaronmobility — Method

polaronmobility(Trange, ε_Inf, ε_S, freq, effectivemass; verbose::Bool=false)

Solves the Feynman polaron problem variationally with finite temperature Osaka energies. From the resulting v, and w parameters, calculates polaron structure (wave function size, etc.). Uses FHIP, Kadanoff (Boltzmann relaxation time) and Hellwarth direct contour integration to predict a temperature-dependent mobility for the material system.

Arguments

Trange::range: temperature range.

ε_Inf: reduced optical dielectric constant.

ε_S: reduced static dielectric constant.

freq: characteristic dielectric phonon frequency (THz).

effectivemass: bare-band effective mass (mₑ).

Returns a structure of type Polaron, containing arrays of useful information. Also prints a lot of information to the standard out - which may be more useful if you're just inquiring as to a particular data point, rather than plotting a temperature-dependent parameter.

As an example, to calculate the electron polaron in MAPI at 300 K:

Examples

polaronmobility(300, 4.5, 24.1, 2.25, 0.12)

PolaronMobility.save_polaron — Method

save_polaron(p::NewPolaron, prefix)

Saves data from 'polaron' into file "prefix". This is a .jdl file for storing the polaron data whilst preserving types. Allows for saving multidimensional arrays that sometimes arise in the polaron data. Each parameter in the NewPolaron type is saved as a dictionary entry. E.g. NewPolaron.α is saved under JLD.load("prefix.jld")["alpha"].

PolaronMobility.savepolaron — Method

savepolaron(fileprefix, p::Polaron)

Saves data from polaron 'p' into file "fileprefix". This is a simple space-delimited text file, with each entry a separate temperature, for plotting with Gnuplot or similar.

Structure of file is written to the header:

Ts, βreds, Kμs, Hμs, FHIPμs, vs, ws, ks, Ms, As, Bs, Cs, Fs, Taus, rfsis

1 2 3 4 5 6 7 8 9 10 11 12 13 14 15

PolaronMobility.κ_i — Method

κ_i(i, v, w)

Calculates the spring-constant coupling the electron to the 'ith' fictitious mass that approximates the exact electron-phonon interaction with a harmonic coupling to a massive fictitious particle.

Required for calculating the polaron free energy.

Note: Not to be confused with the number of physical phonon branches; many phonon branches could be approximated with one or two etc. fictitious masses for example. The number of fictitious mass does not necessarily need to match the number of phonon branches.

Arguments

i::Integer: enumerates the current fictitious mass.

v::Vector{Float64}: is a vector of the v variational parameters.

w::Vector{Float64}: is a vector of the w variational parameters.

PolaronMobility.ϵ_ionic_mode — Method

ϵ_ionic_mode(phonon_mode_freq, ir_activity, volume)

Calculate the ionic contribution to the dielectric function for a given phonon mode.

Arguments

phonon_mode_freq::Float64: is the frequency of the mode in THz.

ir_activity::Float64: is the infra-red activity of the mode in e²amu⁻¹.

volume::Float64: is the volume of the unit cell of the material in m³.

PolaronMobility.ϵ_total — Method

ϵ_total(freqs_and_ir_activity, volume)

Calculate the total ionic contribution to the dielectric function from all phonon modes.

Arguments

freqs_and_ir_activity::Matrix{Float64}: is a matrix containeing the phonon mode frequencies (in THz) in the first column and the infra-red activities (in e²amu⁻¹) in the second column.

volume::Float64: is the volume of the unit cell of the material in m^3.