Functions

PolaronMobility.FrohlichPolaron — Type

Polaron(x...)

Type for storing the polaron information, x....

PolaronMobility.HolsteinPolaron — Type

HolsteinPolaron(x...)

Type for storing the Holstein polaron information, x....

PolaronMobility.Material — Type

mutable struct Material

Structure to contain material-specific data.

PolaronMobility.A — Method

A(v, w, ω, β)

Hellwarth's A expression from Eqn. (62b) in Hellwarth et al. 1999 PRB. Part of the overall free energy expression.

See Hellwarth et a. 1999: https://doi.org/10.1103/PhysRevB.60.299.

PolaronMobility.A — Method

A(v, w, β)

Generalisation of the A function from Eqn. (62b) in Hellwarth et al. 1999. This is the Helmholtz free energy of the trial model.

Required for calculating the polaron free energy.

Arguments

v::Vector{Float64}: is a vector of the v variational parameters.

w::Vector{Float64}: is a vector of the w variational parameters.

β::Union{Float64, Vector{Float64}}: is the reduced thermodynamic temperature ħωⱼ/(kT) associated with the 'jth' phonon mode.

See Hellwarth, R. W., Biaggio, I. (1999): https://doi.org/10.1103/PhysRevB.60.299.

PolaronMobility.A — Method

A(v, w, n)

Calculates A(v, w, β) but at zero-temperature, β = Inf.

Arguments

v::Vector{Float64}: is a vector of the v variational parameters.

w::Vector{Float64}: is a vector of the w variational parameters.

PolaronMobility.C — Method

C(v, w, ω, β)

Hellwarth's C expression from Eqn. (62e) in Hellwarth et al. 1999 PRB. Part of the overall free energy expression.

See Hellwarth et a. 1999: https://doi.org/10.1103/PhysRevB.60.299.

PolaronMobility.C — Method

C(v, w, β)

Generalisation of the C function from Eqn. (62e) in Hellwarth et al. 1999. This is the expected value of the trial action <S_0> taken w.r.t trial action.

Required for calculating the polaron free energy.

Arguments

v::Vector{Float64}: is a vector of the v variational parameters.

w::Vector{Float64}: is a vector of the w variational parameters.

β::Union{Float64, Vector{Float64}}: is the reduced thermodynamic temperature ħωⱼ/(kT) associated with the 'jth' phonon mode.

See Hellwarth, R. W., Biaggio, I. (1999): https://doi.org/10.1103/PhysRevB.60.299.

PolaronMobility.C — Method

C(v, w)

Calculates C(v, w, β) but at zero-temperature, β = Inf.

Arguments

v::Vector{Float64}: is a vector of the v variational parameters.

w::Vector{Float64}: is a vector of the w variational parameters.

PolaronMobility.C_ij — Method

C_ij(i, j, v, w)

Calculates the element to the coupling matrix C_ij (a generalisation of Feynman's C coupling variational parameter in Feynman 1955) between the electron and the ith' andjth' fictitious masses that approximates the exact electron-phonon interaction with a harmonic coupling to a massive fictitious particle.

Required for calculating the polaron free energy.

Note: Not to be confused with the number of physical phonon branches; many phonon branches could be approximated with one or two etc. fictitious masses for example. The number of fictitious mass does not necessarily need to match the number of phonon branches.

Arguments

i::Integer, j::Integer: enumerate the current fictitious masses under focus (also the index of the element in the coupling matrix C)

v::Vector{Float64}: is a vector of the v variational parameters.

w::Vector{Float64}: is a vector of the w variational parameters.

See Feynman 1955: http://dx.doi.org/10.1103/PhysRev.97.660.

PolaronMobility.FHIP_mobility_lowT — Method

FHIP_mobility_lowT(v, w, α, ω, β)

FHIP mobility in the low-temperature approximation.

See also inverse_FHIP_mobility_lowT

PolaronMobility.HellwarthAScheme — Method

HellwarthAScheme(phonon_modes; T=295, convergence=1e-6)

Multiple phonon mode reduction to a single effective frequency. Temperature dependent, defaults to T = 295 K.

Solved iteratively by bisection until Δfreq < convergence.

Follows Hellwarth et al. 1999 PRB 'A' scheme, Eqn. (50) RHS.

See Hellwarth et a. 1999: https://doi.org/10.1103/PhysRevB.60.299.

PolaronMobility.HellwarthBScheme — Method

HellwarthBScheme(LO)

LO an array assumed to be of [freq ; absolute ir activity]

Multiple phonon mode reduction to a single effective frequency. Hellwarth et al. 1999 PRB, 'B scheme'; the athermal method. Averaging procedure is constructed by considering the average effect of the action of multiple branches.

Follows Eqn. (58) in this paper, assuming typo on LHS, should actually be W_e.

See Hellwarth et a. 1999: https://doi.org/10.1103/PhysRevB.60.299.

PolaronMobility.Hellwarth_mobility — Method

Hellwarth_mobility(v, w, α, ω, β)

The Hellwarth polaron mobility.

See also inverse_Hellwarth_mobility

PolaronMobility.Kadanoff_mobility_lowT — Method

Kadanoff_mobility_lowT(v, w, α, ω, β)

Kadanoff mobility in the low-temperature approximation.

See also inverse_Kadanoff_mobility_lowT

PolaronMobility.addunits! — Method

addunits!(polaron::Polaron)

PolaronMobility.addunits! — Method

addunits!(polaron::Material)

PolaronMobility.feynmanvw — Method

feynmanvw(v, w, αωβ...; upper_limit = Inf64)

Minimises the multiple phonon mode free energy function for a set of vₚ and wₚ variational parameters. The variational parameters follow the inequality: v₁ > w₁ > v₂ > w₂ > ... > vₙ > wₙ. Generalises feynmanvw to multiple variational parameters.

Arguments

v::Vector{Float64}: vector of initial v parameters.

w::Vector{Float64}: vector of initial w parameters.

α::Union{Float64, Vector{Float64}}: is the partial dielectric electron-phonon coupling parameter for one or many phonon modes.

ω::Union{Float64, Vector{Float64}}: phonon mode frequencies (2π THz) for one or many phonon modes.

β::Union{Float64, Vector{Float64}}: is the reduced thermodynamic temperature ħωⱼ/(kT) for one or many phonon modes.

See also F.

PolaronMobility.frohlich_complex_conductivity — Method

frohlichcomplexconductivity(Ω, β, α, v, w; rtol = 1e-3)

Calculate the complex conductivity σ(Ω) of the polaron at finite temperatures for a given frequency Ω (equal to 1 / Z(Ω) with Z the complex impedence).

Arguments

Ω::Float64: is the frequency (2π THz) of applied electric field.

β::Float64: is the reduced thermodynamic betas.

α::Float64: is the Frohlich alpha coupling parameter.

v::Float64: is the 'v' variational parameter.

w::Float64: is the 'w' variational parameter.

rtol: relative tolerance for the accuracy of any quadrature integrations.

See also polaron_complex_impedence

PolaronMobility.frohlich_complex_impedence — Method

frohlich_complex_impedence(Ω, β, α, v, w; rtol = 1e-3, T = nothing, verbose = false)

Calculate the complex impedence Z(Ω) of the polaron at finite temperatures for a given frequency Ω (Eqn. (41) in FHIP 1962).

Arguments

Ω::Float64: is the frequency (2π THz) of applied electric field.

β::Float64: is the reduced thermodynamic betas.

α::Float64: is the Frohlich alpha coupling parameter.

v::Float64: is the 'v' variational parameter.

w::Float64: is the 'w' variational parameter.

rtol: relative tolerance for the accuracy of any quadrature integrations.

T: is a token used by make_polaron() to keep track of the temperature for printing during a calculation. Do not alter.

verbose: is used by make_polaron() to specify whether or not to print. Ignore.

See FHIP 1962: https://doi.org/10.1103/PhysRev.127.1004.

PolaronMobility.frohlich_coupling — Method

frohlich_coupling(k, α, ω)

Calculate the coupling strength for the Frohlich continuum polaron model.

Arguments

k: a scalar value representing the k-coordinate in k-space

α: a scalar value representing the coupling constant

ω: a scalar value representing the phonon frequency

Returns

The coupling strength for the Frohlich continuum polaron model.

Example

result = frohlich_coupling(2.0, 0.5, 1.0)
println(result)

Expected Output: 6.0

PolaronMobility.frohlich_energy — Method

F(v, w, α, ω, β)

Calculates the Helmholtz free energy of the polaron for a material with multiple phonon branches.

This generalises the Osaka 1959 (below Eqn. (22)) and Hellwarth. et al 1999 (Eqn. (62a)) free energy expressions.

Arguments

v::Vector{Float64}: is a vector of the v variational parameters.

w::Vector{Float64}: is a vector of the w variational parameters.

α::Union{Float64, Vector{Float64}}: is the partial dielectric electron-phonon coupling parameter for the 'jth' phonon mode.

ω::Union{Float64, Vector{Float64}}: phonon mode frequencies (2π THz).

β::Union{Float64, Vector{Float64}}: is the reduced thermodynamic temperature ħωⱼ/(kT) associated with the 'jth' phonon mode.

See Osaka, Y. (1959): https://doi.org/10.1143/ptp.22.437 and Hellwarth, R. W., Biaggio, I. (1999): https://doi.org/10.1103/PhysRevB.60.299.

PolaronMobility.frohlich_energy_k_space — Method

frohlich_energy_k_space(v, w, α, ωβ...; rₚ = 1, limits = [0, Inf])

Calculate the total energy, kinetic energy, and interaction energy of the Frohlich lattice polaron.

Arguments

v: a scalar value representing a variational paramater.

w: a scalar value representing a variational paramater.

α: a scalar value representing the coupling constant.

ω: a scalar value representing the phonon frequency.

β: (optional) a scalar value representing the inverse temperature.

rₚ: The characteristic polaron radius (default is 1).

limits: The limits of integration for the interaction energy calculation (default is [0, Inf]).

Returns

total_energy: The calculated total polaron energy.

kinetic_energy: The calculated polaron kinetic energy.

interaction_energy: The calculated polaron interaction energy.

PolaronMobility.frohlich_interaction_energy — Method

frohlich_interaction_energy(v, w, α, ωβ...)

Integral of Eqn. (31) in Feynman 1955. Part of the overall ground-state energy expression.

See Feynman 1955: http://dx.doi.org/10.1103/PhysRev.97.660.

PolaronMobility.frohlich_interaction_energy_k_space — Method

frohlich_interaction_energy_k_space(v, w, α, ω, β; rₚ = 1, limits = [0, Inf])

Calculate the Frohlich polaron interaction energy in k-space at finite temperaure.

Arguments

τ: a scalar value representing the imaginary time.

v: a scalar value representing a variational paramater.

w: a scalar value representing a variational paramater.

α: a scalar value representing the coupling constant.

ω: a scalar value representing the phonon frequency.

β: a scalar value representing the inverse temperature.

rₚ: an optional scalar value representing the characteristic polaron radius. Default value is 1.

a: an optional scalar value representing the lattice constant. Default value is 1.

limits: an optional array of two scalar values representing the lower and upper limits of the integration range in k-space. Default value is [-π, π].

Returns

A scalar value representing the Frohlich polaron interaction energy in k-space at finite temperature.

PolaronMobility.frohlich_memory_function_k_space — Method

frohlich_memory_function_k_space(Ω, v, w, α, ω, β; rₚ = 1, limits = [0, Inf])

Calculate the memory function for the Frohlich model in k-space at finite temperature and frequency.

Arguments

Ω: a scalar value representing the frequency at which the memory function is evaluated.

v: a scalar value representing the optimal variational parameter.

w: a scalar value representing the optimal variational parameter.

α: a scalar value representing the coupling constant.

ω: a scalar value representing the phonon frequency.

β: a scalar value representing the inverse temperature.

rₚ: an optional scalar value representing the characteristic polaron radius. Default value is 1.

limits: an array of two scalar values representing the lower and upper limits of the integration range in k-space. Default is an infinite sphere.

Returns

A scalar value representing the memory function of the Frohlich model in k-space at finite temperature evaluated at the frequency Ω.

PolaronMobility.frohlich_mobility — Method

polaron_mobility(v, w, α, ω, β)

The polaron mobility.

See also inverse_polaron_mobility

PolaronMobility.frohlich_mobility_k_space — Method

frohlich_mobility_k_space(v, w, α, ω, β; rₚ = 1, limits = [0, Inf])

Calculate the DC mobility in k-space for a Frohlich polaron system at finite temperature.

Arguments

v: a scalar value representing the optimal variational parameter.

w: a scalar value representing the optimal variational parameter.

α: a scalar value representing the coupling constant.

ω: a scalar value representing the phonon frequency.

β: a scalar value representing the inverse temperature.

rₚ: an optional scalar value representing the characteristic polaron radius. Default value is 1.

limits: an array of two scalar values representing the lower and upper limits of the integration range in k-space. Default is an infinite sphere.

Returns

The DC mobility in k-space for the Frohlich polaron system at finite temperature.

PolaronMobility.frohlich_structure_factor_k_space — Method

frohlich_structure_factor_k_space(t, v, w, α, ωβ...; limits = [0, Inf])

Calculate the structure factor in k-space for the Frohlich continuum polaron model at finite temperature.

Arguments

t: a scalar value representing the real time.

v: a scalar value representing a variational parameter.

w: a scalar value representing a variational parameter.

α: a scalar value representing the coupling constant.

ω: a scalar value representing the phonon frequency.

β: a scalar value representing the inverse temperature.

rₚ: an optional scalar value representing the characteristic polaron radius. Default value is 1.

limits: an array of two scalar values representing the lower and upper limits of the integration range in k-space. Default is an infinite sphere.

Returns

A scalar value representing the calculated structure factor in k-space for the Frohlich continuum polaron model at finite temperature.

PolaronMobility.frohlichalpha — Method

frohlichalpha(ε_Inf, ε_S, freq, m_eff)

Calculates the Frohlich alpha parameter, for a given dielectric constant, frequency (f) of phonon in Hertz, and effective mass (in units of the bare electron mass).

See Feynman 1955: http://dx.doi.org/10.1103/PhysRev.97.660.

PolaronMobility.frohlichalpha — Method

frohlichalpha(ϵ_optic, ϵ_ionic, ϵ_total, phonon_mode_freq, m_eff)

Calculates the partial dielectric electron-phonon coupling parameter for a given longitudinal optical phonon mode.

This decomposes the original Frohlich alpha coupling parameter (defined for a single phonon branch) into contributions from multiple phonon branches.

Arguments

ϵ_optic::Float64: is the optical dielectric constant of the material.

ϵ_ionic::Float64: is the ionic dielectric contribution from the phonon mode.

ϵ_total::Float64: is the total ionic dielectric contribution from all phonon modes of the material.

phonon_mode_freq::Float64: is the frequency of the phonon mode (THz).

m_eff::Float64 is the band mass of the electron (in units of electron mass m_e).

PolaronMobility.frohlichpolaron — Method

polaron(αrange, Trange, Ωrange, ω, v_guesses, w_guesses; verbose=false)

Outer constructor for the Polaron type. This function evaluates model data for the polaron, i.e. unitless and not material specific.

Examples

julia> polaron(6, 300, 3, 1.0, 3.6, 2.8)

PolaronMobility.frohlichpolaron — Method

No frequency input.

PolaronMobility.frohlichpolaron — Method

No temperature input => 300 K.

PolaronMobility.frohlichpolaron — Method

polaron(material::Material, TΩrange...; v_guesses=3.11, w_guesses=2.87, verbose=false)

Material specific constructors that use material specific parameters to parameterise the polaron. Material data is inputted through the Material type. Returns all data in either SI units or other common, suitable units otherwise.

PolaronMobility.frohlichpolaron — Method

Single alpha parameter. polaron() expects alpha parameters to be in a Vector.

PolaronMobility.frohlichpolaron — Method

No input => α = 1

PolaronMobility.h_i — Method

h_i(i, v, w)

Calculates the normal-mode (the eigenmodes) frequency of the coupling between the electron and the `ith' fictitious mass that approximates the exact electron-phonon interaction with a harmonic coupling to a massive fictitious particle.

Required for calculating the polaron free energy.

Note: Not to be confused with the number of physical phonon branches; many phonon branches could be approximated with one or two etc. fictitious masses for example. The number of fictitious mass does not necessarily need to match the number of phonon branches.

Arguments

i::Integer: enumerates the current fictitious mass.

v::Vector{Float64}: is a vector of the v variational parameters.

w::Vector{Float64}: is a vector of the w variational parameters.

PolaronMobility.holstein_coupling — Method

holstein_coupling(k, α, ω; dims = 1)

Calculate the coupling strength for the Holstein lattice polaron model.

Arguments

k: a scalar value representing the k-coordinate in k-space

α: a scalar value representing the coupling constant

ω: a scalar value representing the phonon frequency

dims: an optional scalar value representing the dimensionality of the system. Default value is 3.

Returns

The coupling strength for the Holstein model.

Example

result = holstein_coupling(2.0, 0.5, 1.0; dims = 3)
println(result)

Expected Output: 6.0

PolaronMobility.holstein_energy — Method

Total free energy for the Holstein model. Here the k-space integral is evaluated analytically.

PolaronMobility.holstein_energy_k_space — Method

holstein_energy_k_space(v, w, α, ωβ...; dims = 3, rₚ = 1, a = 1, limits = [-π, π])

Calculate the total energy, kinetic energy, and interaction energy of the Holstein lattice polaron.

Arguments

v: a scalar value representing a variational paramater.

w: a scalar value representing a variational paramater.

α: a scalar value representing the coupling constant.

ω: a scalar value representing the phonon frequency.

β: (optional) a scalar value representing the inverse temperature.

dims: The number of dimensions of the system (default is 3).

rₚ: The characteristic polaron radius (default is 1).

a: The lattice constant (default is 1).

limits: The limits of integration for the interaction energy calculation (default is [-π, π]).

Returns

total_energy: The calculated total polaron energy.

kinetic_energy: The calculated polaron kinetic energy.

interaction_energy: The calculated polaron interaction energy.

PolaronMobility.holstein_interaction_energy — Method

holstein_interaction_energy(v, w, α, ωβ...; dims = 3)

Electron-phonon interaction energy for the Holstein mode at finite temperature. Here the k-space integral is evaluated analytically.

Arguments

v: A scalar representing a variational parameter.

w: A scalar representing a variational parameter.

α: A scalar representing the electron-phonon coupling.

ω: A scalar representing the adiabaticity.

β: A scalar representing the inverse temperature.

dims: An optional argument representing the number of dimensions. Default is 3.

Returns

integral: The electron-phonon interaction energy for the Holstein mode at finite temperature.

Example

v = 0.2
w = 0.1
α = 0.5
ω = 0.3
β = 1.0
result = holstein_interaction_energy(v, w, α, ω, β)
println(result)

This example calculates the electron-phonon interaction energy for given values of v, w, α, ω, and β. The result is then printed.

PolaronMobility.holstein_interaction_energy_k_space — Method

holstein_interaction_energy_k_space(v, w, α, ω, β; dims = 3, rₚ = 1, a = 1, limits = [-π, π])

Calculate the Holstein polaron interaction energy in k-space at finite temperaure.

Arguments

τ: a scalar value representing the imaginary time.

v: a scalar value representing a variational paramater.

w: a scalar value representing a variational paramater.

α: a scalar value representing the coupling constant.

ω: a scalar value representing the phonon frequency.

β: a scalar value representing the inverse temperature.

dims: an optional scalar value representing the dimensionality of the system. Default value is 3.

rₚ: an optional scalar value representing the characteristic polaron radius. Default value is 1.

a: an optional scalar value representing the lattice constant. Default value is 1.

limits: an optional array of two scalar values representing the lower and upper limits of the integration range in k-space. Default value is [-π, π].

Returns

A scalar value representing the Holstein polaron interaction energy in k-space at finite temperature.

PolaronMobility.holstein_memory_function — Method

holstein_memory_function(Ω, v, w, α, ω, β; dims = 3)

Calculate the memory function using the general_memory_function function.

Arguments

Ω: a parameter representing a frequency

v, w, α, ω, β: parameters used to calculate the structure factor

dims: an optional parameter representing the dimensions (default value is 3)

Returns

The result of the general_memory_function function, which represents the calculated memory function.

Example

result = holstein_memory_function(Ω, v, w, α, ω, β; dims = 3)

In this example, the holstein_memory_function is called with the input parameters Ω, v, w, α, ω, and β, and the optional parameter dims set to 3. The function calculates the memory function using the general_memory_function function and returns the result.

PolaronMobility.holstein_memory_function_k_space — Method

holstein_memory_function_k_space(Ω, v, w, α, ωβ...; dims = 3)

Calculate the memory function for the Holstein model in k-space at finite temperature and frequency.

Arguments

Ω: a scalar value representing the frequency at which the memory function is evaluated.

v: a scalar value representing the optimal variational parameter.

w: a scalar value representing the optimal variational parameter.

α: a scalar value representing the coupling constant.

ω: a scalar value representing the phonon frequency.

β: a scalar value representing the inverse temperature.

dims: an optional scalar value representing the dimensionality of the system. Default value is 3.

rₚ: an optional scalar value representing the characteristic polaron radius. Default value is 1.

a: an optional scalar value representing the lattice constant. Default value is 1.

limits: an array of two scalar values representing the lower and upper limits of the integration range in k-space. Default is 1D Brillouin zone.

Returns

A scalar value representing the memory function of the Holstein model in k-space at finite temperature evaluated at the frequency Ω.

PolaronMobility.holstein_mobility — Method

polaron_mobility(v, w, α, ω, β)

The polaron mobility.

See also inverse_polaron_mobility

PolaronMobility.holstein_structure_factor — Method

holstein_structure_factor(t, v, w, α, ω, β; dims = 3)

Calculate the structure factor for the Holstein model.

Arguments

t: A scalar representing the time.

v: A scalar representing a parameter.

w: A scalar representing a parameter.

α: A scalar representing a coupling constant.

ω: A scalar representing a coupling constant.

β: A scalar representing the inverse temperature.

dims: An optional argument representing the number of dimensions. Default is 3.

Returns

The value of the structure factor for the given inputs.

Example

t = 0.5
v = 0.2
w = 0.1
α = 0.3
ω = 0.4
β = 1.0
result = holstein_structure_factor(t, v, w, α, ω, β)
println(result)

This example demonstrates how to use the holstein_structure_factor function to calculate the structure factor for a given set of parameters. The function is called with the values of t, v, w, α, ω, and β as arguments, and the result is then printed.

PolaronMobility.holstein_structure_factor_k_space — Method

holstein_structure_factor_k_space(t, v, w, α, ωβ...; dims = 3)

Calculate the structure factor in k-space for the Holstein lattice polaron model at finite temperature.

Arguments

t: a scalar value representing the real time.

v: a scalar value representing a variational parameter.

w: a scalar value representing a variational parameter.

α: a scalar value representing the coupling constant.

ω: a scalar value representing the phonon frequency.

β: a scalar value representing the inverse temperature.

dims: an optional scalar value representing the dimensionality of the system. Default value is 3.

rₚ: an optional scalar value representing the characteristic polaron radius. Default value is 1.

a: an optional scalar value representing the lattice constant. Default value is 1.

limits: an array of two scalar values representing the lower and upper limits of the integration range in k-space. Default is 1D Brillouin zone.

Returns

A scalar value representing the calculated structure factor in k-space for the Holstein lattice polaron model at finite temperature.

PolaronMobility.holsteinpolaron — Method

No frequency input => Phonon peak.

PolaronMobility.holsteinpolaron — Method

No temperature input => 300 K.

PolaronMobility.holsteinpolaron — Method

holstein(material::Material, TΩrange...; v_guesses=3.11, w_guesses=2.87, verbose=false)

Material specific constructors that use material specific parameters to parameterise the polaron. Material data is inputted through the Material type. Returns all data in either SI units or other common, suitable units otherwise.

PolaronMobility.holsteinpolaron — Method

Single alpha parameter. holstein() expects alpha parameters to be in a Vector.

PolaronMobility.holsteinpolaron — Method

No input => α = 1

PolaronMobility.inverse_FHIP_mobility_lowT — Method

inverse_FHIP_mobility_lowT(v, w, α, ω, β)

FHIP low-temperature mobility, final result of Feynman1962. [1.60] in Devreese2016 page 36; 6th Edition of Frohlich polaron notes (ArXiv).

See also FHIP_mobility_lowT

PolaronMobility.inverse_FHIP_mobility_lowT — Method

inverse_FHIP_mobility_lowT(v, w, α::Vector, ω::Vector, β::Vector)

Inverse FHIP mobility for multiple phonon modes.

See also FHIP_mobility_lowT

PolaronMobility.inverse_Hellwarth_mobility — Method

inverse_Hellwarth_mobility(v, w, α, ω, β)

Calculates the DC mobility using Hellwarth et al. 1999 Eqn. (2). Directly performs contour integration in Feynman1962, for finite temperature DC mobility. Eqns. (2) and (1) are going back to the general (pre low-T limit) formulas in Feynman1962. To evaluate these, you need to do the explicit contour integration to get the polaron self-energy.

See Hellwarth et a. 1999: https://doi.org/10.1103/PhysRevB.60.299.

See also Hellwarth_mobility

PolaronMobility.inverse_Hellwarth_mobility — Method

inverse_Hellwarth_mobility(v, w, α::Vector, ω::Vector, β::Vector)

Inverse Hellwarth mobility for multiple phonon modes.

See also Hellwarth_mobility

PolaronMobility.inverse_Kadanoff_mobility_lowT — Method

inverse_Kadanoff_mobility_lowT(v, w, α, ω, β)

Kadanoff low-temperaure mobility, constructed around Boltzmann equation. Adds factor of 3 / (2 * β) c.f. FHIP, correcting phonon emission behaviour.

See also Kadanoff_mobility_lowT

PolaronMobility.inverse_Kadanoff_mobility_lowT — Method

inverse_Kadanoff_mobility_lowT(v, w, α::Vector, ω::Vector, β::Vector)

Inverse Kadanoff mobility for multiple phonon modes.

See also Kadanoff_mobility_lowT

PolaronMobility.inverse_frohlich_mobility — Method

inverse_polaron_mobility(v, w, α, ω, β)

Calculate the inverse of the dc mobility μ of the polaron at finite temperatues (Eqn. (11.5) in F. Peeters and J. Devreese 1984) for a given frequency Ω.

Arguments

v::Float64: is the 'v' variational parameter.

w::Float64: is the 'w' variational parameter.

α::Float64: is the Frohlich alpha coupling parameter.

ω::Float64: is the angular phonon frequency.

β::Float64: is the reduced thermodynamic beta.

verbose: is used by make_polaron() to specify whether or not to print. Ignore.

See F. Peeters and J. Devreese 1984: https://doi.org/10.1016/S0081-1947(08)60312-4.

See also polaron_mobility, polaron_complex_conductivity

PolaronMobility.inverse_frohlich_mobility — Method

inverse_polaron_mobility(v, w, α::Vector, ω::Vector, β::Vector)

inverse of the polaron mobility, but for multiple phonon modes.

PolaronMobility.inverse_holstein_mobility — Method

holstein_mobility(v, w, α, ω, β; dims = 3)

Calculate the mobility using the general_memory_function and holstein_structure_factor functions.

Arguments

v: Parameter used in the holstein_structure_factor function.

w: Parameter used in the holstein_structure_factor function.

α: Parameter used in the holstein_structure_factor function.

ω: Parameter used in the holstein_structure_factor function.

β: Parameter used in the holstein_structure_factor function.

dims: The dimensionality of the system. Default is 3.

Returns

The calculated mobility.

Example

v = 1.0
w = 2.0
α = 0.5
ω = 1.0
β = 0.2
dims = 3

result = holstein_mobility(v, w, α, ω, β, dims=dims)
println(result)

This code calculates the mobility using the given parameters and prints the result.

PolaronMobility.inverse_holstein_mobility_k_space — Method

holstein_mobility_k_space(v, w, α, ω, β; dims = 3, rₚ = 1, a = 1, limits = [-π, π])

Calculate the DC mobility in k-space for a Holstein polaron system at finite temperature.

Arguments

v: a scalar value representing the optimal variational parameter.

w: a scalar value representing the optimal variational parameter.

α: a scalar value representing the coupling constant.

ω: a scalar value representing the phonon frequency.

β: a scalar value representing the inverse temperature.

dims: an optional scalar value representing the dimensionality of the system. Default value is 3.

rₚ: an optional scalar value representing the characteristic polaron radius. Default value is 1.

a: an optional scalar value representing the lattice constant. Default value is 1.

limits: an array of two scalar values representing the lower and upper limits of the integration range in k-space. Default is 1D Brillouin zone.

Returns

The DC mobility in k-space for the Holstein polaron system at finite temperature.

PolaronMobility.load_frohlich_polaron — Method

load_polaron(p::FrohlichPolaron, prefix)

Loads data from file "polaronfilepath" into a NewPolaron type.

PolaronMobility.load_holstein_polaron — Method

load_polaron(p::NewPolaron, prefix)

Loads data from file "polaronfilepath" into a NewPolaron type.

PolaronMobility.material — Method

material(ϵ_optical, ϵ_static, m_eff, phonon_freq)

Construct a 'struct Material' object from traditional Frohlich Hamiltonian parameters. 

ϵ_ionic and α derived from these parmaters; other material properties filled with bogus values.

PolaronMobility.material — Method

material(ϵ_optic, ϵ_static, m_eff, phonon_freqs, ir_activity, volume)

Construct a 'struct Material' object from a set of infrared activities and phonon
frequencies, for use with the 'multiple phonon branches' extension of the code. 

effective_freq calculated with the Hellwarth 'B' scheme; ϵ_ionic and α calculated from
the infrared activties.

PolaronMobility.phonon_propagator — Method

phonon_propagator(τ, ω, β)

Calculate the value of the phonon propagator at a given imaginary time (τ), phonon frequency (ω), and inverse temperature (β).

Arguments

τ: A scalar representing the imaginary time.

ω: A scalar representing the phonon frequency.

β: A scalar representing the inverse temperature.

Returns

The value of the phonon propagator.

Example

τ = 0.5
ω = 0.2
β = 1.0
result = phonon_propagator(τ, ω, β)
println(result)

This example calculates the value of the phonon propagator for τ = 0.5, ω = 0.2, and β = 1.0. The result is then printed.

PolaronMobility.phonon_propagator — Method

phonon_propagator(τ, ω)

Calculate the value of the phonon propagator based on the given inputs.

Arguments

τ: A scalar representing the imaginary time.

ω: A scalar representing the adiabaticity.

Example

τ = 0.5
ω = 0.2
result = phonon_propagator(τ, ω)
println(result)

This example calculates the value of the phonon propagator for the given values of τ and ω. The result is then printed.

Returns

The value of the phonon propagator.

PolaronMobility.polaron_memory_function — Method

polaron_memory_function(Ω, structure_factor; limits = [0, Inf])

This function calculates the integral of a given structure factor with respect to time using the quadgk function from the QuadGK package.

Arguments

Ω: A scalar representing the frequency.

structure_factor: A function that returns the value of the structure factor at a given time.

limits (optional): A 2-element array specifying the lower and upper limits of integration. Default is [0, Inf].

Returns

The integral of the structure factor with respect to time.

Example

# Define a structure factor function
function structure_factor(t)
    # Implementation of the structure factor
    # ...
end

# Calculate the memory function for a given frequency and structure factor
Ω = 0.5
limits = [0, 10]
result = general_memory_function(Ω, structure_factor; limits = limits)
println(result)

This example demonstrates how to use the general_memory_function to calculate the memory function for a given frequency Ω and structure factor function structure_factor. The limits argument is optional and specifies the lower and upper limits of integration. The result is then printed.

PolaronMobility.polaron_memory_function — Method

general_memory_function(structure_factor; limits = [0, Inf])

Calculate the integral of a given function structure_factor using the quadgk function in Julia.

Arguments

structure_factor: A function that takes a single argument t and returns a value.

limits: An optional array specifying the lower and upper limits of integration. Default is [0, Inf].

Returnstyp

integral: The result of the numerical integration of the function structure_factor over the specified limits.

Example

# Define the structure factor function
function structure_factor(t)
    return t^2 + 2t + 1
end

# Call the general_memory_function with the structure_factor function
result = general_memory_function(structure_factor; limits = [0, 10])

println(result)  # Output: 383.3333333333333

PolaronMobility.polaron_propagator — Method

polaron_propagator(τ, v, w)

Calculate the value of the polaron propagator based on the given inputs.

Arguments

τ::Number: A scalar representing the imaginary time.

v::Number: A scalar representing a variational parameter.

w::Number: A scalar representing a variational parameter.

Returns

The value of the polaron propagator.

Example

τ = 0.5
v = 0.2
w = 0.1
result = polaron_propagator(τ, v, w)
println(result)

This example calculates the polaron propagator for the given values of τ, v, and w. The result is then printed.

PolaronMobility.polaron_propagator — Method

polaron_propagator(τ, v, w, β)

Calculate the imaginary time polaron Green's function with temperature dependence.

Arguments

τ: A scalar representing the imaginary time.

v: A scalar representing a variational parameter.

w: A scalar representing a variational parameter.

β: A scalar representing the inverse temperature.

Returns

The value of the polaron propagator.

Example

τ = 0.5
v = 0.2
w = 0.1
β = 1.0
result = polaron_propagator(τ, v, w, β)
println(result)

This example calculates the polaron propagator for given values of τ, v, w, and β. The result is then printed.

PolaronMobility.polaron_propagator — Method

D(τ, v, w, β)

Calculates the recoil function (a generalisation of D(u) in Eqn. (35c) in FHIP 1962) that approximates the exact influence (recoil effects) of the phonon bath on the electron with the influence of the fictitious masses attached by springs to the electron. It appears in the exponent of the intermediate scattering function.

Arguments

τ::Float64: is the imaginary time variable.

v::Vector{Float64}: is a vector of the v variational parameters.

w::Vector{Float64}: is a vector of the w variational parameters.

β::Union{Float64, Vector{Float64}}: is the reduced thermodynamic temperature ħωⱼ/(kT) associated with the 'jth' phonon mode.

See FHIP 1962: https://doi.org/10.1103/PhysRev.127.1004.

PolaronMobility.polaron_propagator — Method

D(τ, v, w)

Calculates the recoil function at zero-temperature.

Arguments

τ::Float64: is the imaginary time variable.

v::Vector{Float64}: is a vector of the v variational parameters.

w::Vector{Float64}: is a vector of the w variational parameters.

PolaronMobility.puconvert — Method

puconvert(x::Unitful.Quantity)

Convert a quantity to the appropriate polaron unit.

PolaronMobility.puconvert — Method

puconvert(u::Unitful.Units, x::Number)

Interpret x as a quantity given in polaron units and convert it to the unit u.

PolaronMobility.punit — Method

punit(x::Unitful.Quantity)
punit(x::Unitful.Units)
punit(x::Unitful.Dimensions)

Returns the appropriate polaron unit (a Unitful.Units object) for the dimension of x.

PolaronMobility.pustrip — Method

pustrip(x::Unitful.Quantity)

Returns the value of the quantity converted to polaron units as a number type (i.e., with the units removed). This is equivalent to Unitful.ustrip(puconvert(x)).

PolaronMobility.save_frohlich_polaron — Method

save_polaron(p::FrohlichPolaron, prefix)

Saves data from 'polaron' into file "prefix". This is a .jdl file for storing the polaron data whilst preserving types. Allows for saving multidimensional arrays that sometimes arise in the polaron data. Each parameter in the NewPolaron type is saved as a dictionary entry. E.g. NewPolaron.α is saved under JLD.load("prefix.jld")["alpha"].

PolaronMobility.save_holstein_polaron — Method

save_polaron(p::Polaron, prefix)

Saves data from 'polaron' into file "prefix". This is a .jdl file for storing the polaron data whilst preserving types. Allows for saving multidimensional arrays that sometimes arise in the polaron data. Each parameter in the NewPolaron type is saved as a dictionary entry. E.g. NewPolaron.α is saved under JLD.load("prefix.jld")["alpha"].

PolaronMobility.trial_energy — Method

trial_energy(v, w, ω, β)

Calculate the free electron energy at finite temperature.

Arguments

v: a scalar value representing a variational paramater.

w: a scalar value representing a variational paramater.

ω: a scalar value representing the phonon frequency.

β: a scalar value representing the inverse temperature.

Returns

A scalar value representing the calculated free electron energy at finite temperature.

Example

v = 0.5
w = 1.0
ω = 2.0
β = 0.2
result = electron_energy(v, w, ω, β; dims = 3)
println(result)

Expected Output: A scalar value representing the calculated free electron energy at finite temperature.

PolaronMobility.vw_variation — Method

vw_variation(energy, initial_v, initial_w; lower_bounds = [0, 0], upper_bounds = [Inf, Inf])

This function optimizes the values of v and w to minimize the energy function energy(x[1] + x[2], x[2])[1]. It uses the Optim package to perform the optimization and returns the optimized values of v and w, as well as the minimized energy, kinetic energy, and potential energy.

Arguments

energy: A function that takes two arguments x and y and returns an array of energy components.

initial_v: The initial value of v.

initial_w: The initial value of w.

lower_bounds: An optional array representing the lower bounds for v and w optimization. Default is [0, 0].

upper_bounds: An optional array representing the upper bounds for v and w optimization. Default is [Inf, Inf].

Returns

Δv + w: The optimized value of v.

w: The optimized value of w.

energy_minimized: The minimized energy.

kinetic_energy: The kinetic energy corresponding to the minimized energy.

potential_energy: The potential energy corresponding to the minimized energy.

Example

energy(x, y) = [x^2 + y^2, x^2, y^2]
initial_v = 0.5
initial_w = 0.2
lower_bounds = [0, 0]
upper_bounds = [Inf, Inf]
result = vw_variation(energy, initial_v, initial_w; lower_bounds, upper_bounds)
println(result)

This example demonstrates how to use the vw_variation function. It defines an energy function energy(x, y) that returns an array of energy components. It then calls vw_variation with the initial values of v and w, as well as lower and upper bounds for the optimization. The function optimizes v and w to minimize the energy and returns the optimized values, as well as the minimized energy, kinetic energy, and potential energy. The result is then printed.

PolaronMobility.κ_i — Method

κ_i(i, v, w)

Calculates the spring-constant coupling the electron to the 'ith' fictitious mass that approximates the exact electron-phonon interaction with a harmonic coupling to a massive fictitious particle.

Required for calculating the polaron free energy.

Note: Not to be confused with the number of physical phonon branches; many phonon branches could be approximated with one or two etc. fictitious masses for example. The number of fictitious mass does not necessarily need to match the number of phonon branches.

Arguments

i::Integer: enumerates the current fictitious mass.

v::Vector{Float64}: is a vector of the v variational parameters.

w::Vector{Float64}: is a vector of the w variational parameters.

PolaronMobility.ϵ_ionic_mode — Method

ϵ_ionic_mode(phonon_mode_freq, ir_activity, volume)

Calculate the ionic contribution to the dielectric function for a given phonon mode.

Arguments

phonon_mode_freq::Float64: is the frequency of the mode in THz.

ir_activity::Float64: is the infra-red activity of the mode in e²amu⁻¹.

volume::Float64: is the volume of the unit cell of the material in m³.

PolaronMobility.ϵ_total — Method

ϵ_total(freqs_and_ir_activity, volume)

Calculate the total ionic contribution to the dielectric function from all phonon modes.

Arguments

freqs_and_ir_activity::Matrix{Float64}: is a matrix containeing the phonon mode frequencies (in THz) in the first column and the infra-red activities (in e²amu⁻¹) in the second column.

volume::Float64: is the volume of the unit cell of the material in m^3.