Functions

PolaronMobility.frohlichalpha — Method

frohlichalpha(εInf,εS,freq,m_eff)

Calculates the Frohlich alpha parameter, for a given dielectric constant,
frequency (f) of phonon in Hertz, and effective mass (in units of the
bare electron mass).

See Feynman 1955:
http://dx.doi.org/10.1103/PhysRev.97.660

PolaronMobility.feynmanvw — Method

feynmanvw(α; v = 0.0, w = 0.0)

Calculate v and w variational polaron parameters,
for the supplied α Frohlich coupling.
This version uses the original athermal action (Feynman 1955).
Returns v,w.

PolaronMobility.polaronmobility — Method

polaronmobility(Trange, εInf, εS, freq, effectivemass; verbose::Bool=false)

Solves the Feynman polaron problem variationally with finite temperature
Osaka energies.  From the resulting v, and w parameters, calculates polaron
structure (wave function size, etc.).  Uses FHIP, Kadanoff (Boltzmann
relaxation time) and Hellwarth direct contour integration to predict
a temperature-dependent mobility for the material system.
Input is a temperature range (e.g. 10:50:1000),
reduced dielectric constants (e.g. 5, 20),
characteristic dielectric phonon frequency (e.g. 2.25E12) - units Hertz
bare-band effective-mass (e.g. 012) - units electron mass.

Returns a structure of type Polaron, containing arrays of useful
information.  Also prints a lot of information to the standard out - which
may be more useful if you're just inquiring as to a particular data point,
rather than plotting a temperature-dependent parameter.

As an example, to calculate the electron polaron in MAPI at 300 K:
polaronmobility(300, 4.5, 24.1, 2.25E12, 0.12)

PolaronMobility.HellwarthBScheme — Method

HellwarthBScheme(LO)

Multiple phonon mode reduction to a single effective frequency.
Hellwarth et al. 1999 PRB, 'B scheme'; the athermal method.
Averaging procedure is constructed by considering the average effect of the action of multiple branches.

Follows Eqn (58) in this paper, assuming typo on LHS, should actually be W_e.

Missing docstring.

Missing docstring for HellwarthAScheme(LO,T=295). Check Documenter's build log for details.

PolaronMobility.savepolaron — Method

savepolaron(fileprefix, p::Polaron)

Saves data from polaron 'p' into file "fileprefix". This is a simple space-delimited text file, with each entry a separate temperature, for plotting with Gnuplot or similar.

Structure of file is written to the header:

Ts, βreds, Kμs, Hμs, FHIPμs, vs, ws, ks, Ms, As, Bs, Cs, Fs, Taus, rfsis

1 2 3 4 5 6 7 8 9 10 11 12 13 14 15

Missing docstring.

Missing docstring for ImX(nurange,v,w,βred,α,ω,mb). Check Documenter's build log for details.